Conflicts:
	WindowsAzurePowershell/src/Commands.ScenarioTest/Commands.ScenarioTest.csproj

+++ WindowsAzurePowershell/tools/BuildInstaller.ps1
@@ -15,7 +15,40 @@
 $scriptFolder = Split-Path -Path $MyInvocation.MyCommand.Definition -Parent
 . ($scriptFolder + '.\SetupEnv.ps1')
 
-Remove-Item -Path "$env:AzurePSRoot\..\Package" -Force -Recurse
+$packageFolder="$env:AzurePSRoot\..\Package"
+if (Test-Path $packageFolder) {
+    Remove-Item -Path "$env:AzurePSRoot\..\Package" -Force -Recurse	
+}
+
+$keyPath = "HKLM:\SOFTWARE\Microsoft\Windows Installer XML"
+if (${env:ADX64Platform}){
+    $keyPath = "HKLM:\SOFTWARE\Wow6432Node\Microsoft\Windows Installer XML"
+}
+
+$allWixVersions = Get-ChildItem $keyPath
+if ($allWixVersions -ne $null){
+    foreach ($wixVersion in $allWixVersions){
+        $wixInstallRoot = $product.GetValue("InstallRoot", $null)
+        if ($wixInstallRoot -ne $null) {
+            Write-Verbose "WIX tools was installed at $wixInstallRoot"
+            break
+        }
+    }
+}
+
+if ($wixInstallRoot -eq $null){
+     Write-Host "You don't have Windows Installer XML Toolset installed, which is needed to build setup." -ForegroundColor "Yellow"
+     Write-Host "Press (Y) to install through codeplex web page we will open for you; (N) to skip"    
+     $keyPressed = $host.UI.RawUI.ReadKey("NoEcho,IncludeKeyUp")
+     if ($keyPressed.Character -eq "y" ){
+        Invoke-Expression "cmd.exe /C start http://wix.codeplex.com/downloads/get/762937"
+        Read-Host "Press any key to continue after the installtion is finished"
+     }
+}
+
+#add wix to the PATH. Note, no need to be very accurate here, 
+#and we just register both 3.8 & 3.5 to simplify the script
+$env:path = $env:path + ";$wixInstallRoot"
 
 # Build the cmdlets in debug mode
 msbuild "$env:AzurePSRoot\..\build.proj" /t:"BuildDebug"
@@ -26,4 +59,4 @@ msbuild "$env:AzurePSRoot\..\build.proj" /t:"BuildDebug"
 # Build the installer
 msbuild "$env:AzurePSRoot\..\build.proj" /t:"BuildSetupDebug"
 
-Write-Verbose "MSI file path: $env:AzurePSRoot\setup\build\Debug\x86\windowsazure-powershell.msi"
+Write-Host "MSI file path: $env:AzurePSRoot\setup\build\Debug\x86\windowsazure-powershell.msi"
